NEF-1677 | Пределал компонент tble в ts и поправил стили
